refactor(theme): to hide code block header when title is not defined, copy button without tooltip
