+++ packages/gatsby-theme-docs/package.json
@@ -19,7 +19,6 @@
   },
   "main": "index.js",
   "dependencies": {
-    "@commercetools-frontend/ui-kit": "10.11.0",
     "@mdx-js/mdx": "1.5.1",
     "@mdx-js/react": "1.5.1",
     "@sentry/browser": "5.9.1",
@@ -45,18 +44,15 @@
     "hast-util-heading": "1.0.3",
     "lodash.escape": "4.0.1",
     "lodash.throttle": "4.1.1",
-    "moment": "2.24.0",
-    "moment-timezone": "0.5.27",
     "parse-numeric-range": "0.0.2",
     "parse5": "5.1.1",
     "prism-themes": "1.3.0",
     "prismjs": "1.17.1",
     "prop-types": "15.7.2",
     "react-helmet": "5.2.1",
-    "react-intl": "3.6.2",
-    "react-router-dom": "5.1.2",
     "rehype-slug": "2.0.3",
-    "remark-emoji": "2.0.2"
+    "remark-emoji": "2.0.2",
+    "tiny-invariant": "1.0.6"
   },
   "devDependencies": {
     "gatsby": "2.18.2",
