+++ packages/gatsby-theme-docs/src/design-system.js
@@ -20,6 +20,7 @@ const uikitColors = {
   colorNeutral90: 'hsl(0, 0%, 90%)',
   colorNeutral95: 'hsl(0, 0%, 95%)',
   colorAccent: '#213c45',
+  colorAccent30: 'hsl(195, 35.2941176471%, 30%)',
   colorAccent40: 'hsl(195, 35.2941176471%, 40%)',
   colorAccent95: 'hsl(195, 35.2941176471%, 95%)',
   colorAccent98: 'hsl(195, 35.2941176471%, 98%)',
@@ -42,6 +43,7 @@ export const colors = {
     surfaceSecondary2: uikitColors.colorNeutral90,
     surfaceSecondary3: uikitColors.colorNeutral60,
     surfaceCode: uikitColors.colorAccent,
+    surfaceCodeCopy: uikitColors.colorAccent30,
     surfaceCodeHighlight: uikitColors.colorAccent40,
     surfaceInfo: uikitColors.colorInfo95,
     surfaceBeta: uikitColors.colorInfo85,
