+++ websites/docs-smoke-test/src/content/components/code-blocks.mdx
@@ -3,5 +3,9 @@ title: Code Blocks
 ---
 
 ```
-Look, _this_ is a **code block**!
+Look, _this_ is a **code block** without a title!
+```
+
+```md title="The title of the code block"
+_This_ is a **code block** with a title
 ```
